python_analytics/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.5
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==1.24.4
packaging==24.0
pandas==2.3.1
pillow==10.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.0
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
streamlit==1.35.0
threadpoolctl==3.6.0
tzdata==2025.2
python-dotenv
